extract addLayer functions
