+++ app/javascript/component-lib/components/map/VMap.vue
@@ -21,12 +21,13 @@ import * as turf from "@turf/turf"
 
 import LayersControl from "./helpers/layers-control.js"
 import { getFirstSymbolLayerId, correctTabFlow } from "./helpers/map-helpers.js"
-import { mixinCarto } from "./mixins/mixin-carto.js"
+import mixinCarto from "./mixins/mixin-carto.js"
+import mixinAddLayers from "./mixins/mixin-add-layers.js"
 import { eventHub } from "../../../vue.js"
 import FilterPane from "./filters/FilterPane"
 
 export default {
-  mixins: [mixinCarto],
+  mixins: [mixinCarto, mixinAddLayers],
 
   components: {
     FilterPane
@@ -95,78 +96,10 @@ export default {
 
     addLayer(layer) {
       if (layer.type === "Raster") {
-        this.addRasterLayer(layer)
+        this.addRasterLayer(layer, this.firstSymbolLayerId)
       } else {
-        this.addVectorLayer(layer)
+        this.addVectorLayer(layer, this.firstSymbolLayerId)
       }
-    },
-
-    addRasterLayer(layer) {
-      const sourceOptions = {
-        type: "raster",
-        tileSize: 256
-      }
-      // Mapbox tileset
-      if (layer.mapbox && layer.mapbox.tileset) {
-        sourceOptions.url = `mapbox://${layer.mapbox.tileset}`
-      // Generic tile endpoint
-      } else if (layer.mapbox && layer.mapbox.endpoint) {
-        sourceOptions.tiles = [layer.mapbox.endpoint]
-      }
-
-      this.map.addLayer({
-        id: layer.name,
-        type: "raster",
-        source: sourceOptions,
-        paint: {
-          "raster-resampling": "nearest",
-          "raster-opacity": 0.5
-        },
-        layout: {
-          visibility: layer.visible ? "visible" : "none"
-        }
-      }, this.firstSymbolLayerId)
-    },
-
-    addVectorLayer(layer) {
-      const carto = layer.carto
-      const tiles = this.createCartoTiles(carto)
-      const isLineType = layer.type === "VectorLine"
-
-      tiles.getTiles(() => {
-        const options = {
-          id: carto.id,
-          type: isLineType ? "line" : "fill",
-          source: {
-            type: "vector",
-            tiles: tiles.mapProperties.mapProperties.metadata.tilejson.vector.tiles
-          },
-          "source-layer": "layer0",
-          layout: {
-            visibility: layer.visible ? "visible" : "none"
-          }
-        }
-
-        if (isLineType) {
-          options.paint = {
-            "line-width": 3,
-            "line-color": carto.colour
-          }
-        } else {
-          options.paint = {
-            "fill-color": carto.colour
-              ? carto.colour
-              : `#${(Math.random().toString(16) + "000000").substring(2, 8)}`,
-            "fill-opacity": 0.5
-          }
-        }
-        
-        this.map.addLayer(options, this.firstSymbolLayerId)
-      })
-    },
-
-    createCartoTiles(carto) {
-      return this.createTiles(this.cartoUsername, this.cartoApiKey, carto)
     },
 
     setLayers(layerSet) {
